Added the ability to update the members of a group via Group Webs Service. As of now, visiting the url as specified below will update the group membership information: /uw_gws/groups/<GROUP>/members/update/
This url will update the django group memberships and provide the following features:
If the group doesn't exist in the db, it will be created.
Users who were found in the Groups Web Service and weren't in the django db will be created upon updating.
Users who are no longer part of a group will be removed from the group. These users will still remain in the auth.users model for django.
The template associated with this view lists out the details of the update, such as users who were created, removed, or unaffected.